Scope mail form styles to component.
I think this is a bug, but I can't get the svelte REPL working to make
an example. The "p, label" css in the MailSignupForm component is
affecting the paragraph tag in the header page (the one that says
"Cory's Blog") in pages that have the mail sign up form embedded. Not
sure how this is happening, because component styles are supposed to be
scoped to the component only?