Implement http to https redirect.
See here:
https://help.heroku.com/J2R1S4T8/can-heroku-force-an-application-to-use-ssl-tls

here:
https://kit.svelte.dev/docs/hooks#handle

and here:
https://stackoverflow.com/questions/71360354/redirecting-from-exported-handle-hook-in-sveltekit

The object is that the Heroku router will set X-Forwarded-Proto in the
request header and leaves it up to the application to check if this
value is not 'https' (is 'http' or something) and then redirect.

The second link shows how to use hooks that run on every request to
do the check and then the third link is an example of how to use
SvelteKit hooks to do a redirect.