Remove debug message. https redirect works I guess?
I don't understand completely what is happening. The heroku router will
receive an http request but then when I print out the url parameter in
sveltekit, the protocol and url will point to "https". So there's no
need to do the "replace()" call, just redirecting to the provided url
will do the trick.

Note, however, that Chrome still acts weird and if you send it an http
request, on my computer, it pops up an xdg-open dialog box and then
doesn't perform any action. On firefox and curl, this works as you would
expect, doing a redirect and serving the https version.

Man, fuck web development.