Remove auth from registerEmail endpoint.
Since this is done by submitting a key in the POST headers, this is
visible to the end-user, exposing the API secret key to the client in
the process.

This is not the place to use basic HTTP auth to secure an endpoint in
this manner because the user doesn't need to enter in credentials.
Since the form is publically accessible, the API endpoint naturally ends
up being exposed in the same manner.

What you could do is at rate limiting (this is needed anyway, but on the
TODO list). Also, consider whitelisting this endpoint to only allow
requests originating from the specified domain(s)?