Remove svelte:head component from layout.
The title tag from individual pages is being replaced by the title tag
value from the __layout.svelte page. This is only manifesting in the
prod server and cannot be reproduced locally. What appears to be
happening is that the base page is loaded and then there is a redirect
happening somewhere. Not sure why this is only happening to the prod but
could be because of DNS CNAME record?